test/Nosql-database/app/requirements.txt
pymongo
paho-mqtt==2.1.0
PyJWT==2.8.0
PyYAML==6.0.1